Use Blueprints to modularize the app
Now is more clear to see wich are the concerns of the application untill now (`hello` and `welcome`)
