+++ flask_tech_talk/hello/__init__.py
@@ -1,0 +1,7 @@
+from flask import Blueprint
+
+
+bp_hello = Blueprint('hello', __name__)
+
+
+from .routes import *
