+++ flask_tech_talk/welcome/__init__.py
@@ -1,0 +1,7 @@
+from flask import Blueprint
+
+
+bp_welcome = Blueprint('welcome', __name__)
+
+
+from .routes import *
